+++ core/examples/extract_dpe.py
@@ -4,14 +4,14 @@ from pipelex.core.working_memory_factory import WorkingMemoryFactory
 from pipelex.pipelex import Pipelex
 from pipelex.pipeline.execute import execute_pipeline
 
-from pipelex_libraries.pipelines.examples.power_extractor.power_extractor import Dpe
+from pipelex_libraries.pipelines.examples.extract_dpe.extract_dpe import Dpe
 from utils.results_utils import get_results_dir_path
 
 SAMPLE_NAME = "pdf_power_extractor_dpe"
 PDF_PATH = "assets/extract_dpe/dpe_single_page.pdf"
 
 
-async def power_extractor(pdf_url: str) -> Dpe:
+async def extract_dpe(pdf_url: str) -> Dpe:
     working_memory = WorkingMemoryFactory.make_from_pdf(
         pdf_url=pdf_url,
         concept_code="documents.PDF",
@@ -29,7 +29,7 @@ async def power_extractor(pdf_url: str) -> Dpe:
 # start Pipelex
 Pipelex.make()
 # run sample using asyncio
-dpe = asyncio.run(power_extractor(pdf_url=PDF_PATH))
+dpe = asyncio.run(extract_dpe(pdf_url=PDF_PATH))
 
 # output results
 output_dir = get_results_dir_path(sample_name=SAMPLE_NAME)
