+++ core/examples/extract_proof_of_purchase.py
@@ -4,14 +4,14 @@ from pipelex.core.working_memory_factory import WorkingMemoryFactory
 from pipelex.pipelex import Pipelex
 from pipelex.pipeline.execute import execute_pipeline
 
-from pipelex_libraries.pipelines.examples.power_extractor.power_extractor import ProofOfPurchase
+from pipelex_libraries.pipelines.examples.extract_proof_of_purchase.models import ProofOfPurchase
 from utils.results_utils import get_results_dir_path
 
 SAMPLE_NAME = "pdf_power_extractor_proof_of_purchase"
 PDF_PATH = "assets/extract_proof_of_purchase/restaurant_invoice.pdf"
 
 
-async def power_extractor(pdf_url: str) -> ProofOfPurchase:
+async def extract_proof_of_purchase(pdf_url: str) -> ProofOfPurchase:
     working_memory = WorkingMemoryFactory.make_from_pdf(
         pdf_url=pdf_url,
         concept_code="documents.PDF",
@@ -29,7 +29,7 @@ async def power_extractor(pdf_url: str) -> ProofOfPurchase:
 # start Pipelex
 Pipelex.make()
 # run sample using asyncio
-proof_of_purchase = asyncio.run(power_extractor(pdf_url=PDF_PATH))
+proof_of_purchase = asyncio.run(extract_proof_of_purchase(pdf_url=PDF_PATH))
 
 # output results
 output_dir = get_results_dir_path(sample_name=SAMPLE_NAME)
