--- pipelex_libraries/pipelines/examples/expense_report/employee.toml
@@ -1,18 +1,0 @@
-
-
-domain = "expense_report"
-
-[concept]
-Employee = "Employee information for expense tracking"
-
-[pipe]
-[pipe.extract_employee_from_expense_report]
-PipeLLM = "Extract employee information from expense report"
-inputs = { expense_report_text = "ExpenseReportText" }
-output = "Employee"
-system_prompt = """You are an expert at extracting employee information from expense reports."""
-prompt_template = """
-Extract employee information from this text:
-
-@expense_report_text
-"""
