--- pipelex_libraries/pipelines/examples/expense_report/expense.toml
@@ -1,24 +1,0 @@
-
-
-domain = "expense_report"
-
-[concept]
-Expense = "A single expense line item/article"
-
-[pipe]
-[pipe.extract_expenses_from_expense_report]
-PipeLLM = "Extract the list of expenses from an expense report"
-inputs = { expense_report_text = "ExpenseReportText" }
-output = "Expense"
-multiple_output = true
-system_prompt = """You are an expert at extracting business expenses from invoices and expense reports.
-You must split combined expenses into individual line items."""
-prompt_template = """
-Extract all expenses from this expense report:
-
-@expense_report_text
-
-Important:
-- Split combined expenses into individual line items
-- Each expense should represent a single article/item
-"""
