--- pipelex_libraries/pipelines/examples/expense_report/vat.toml
@@ -1,30 +1,0 @@
-
-
-domain = "expense_report"
-
-[pipe.calculate_vat]
-PipeLLM = "Calculate VAT for the expense"
-inputs = { expense = "Expense" }
-output = "VAT"
-system_prompt = """
-You are an expert in French VAT rules for business expenses.
-Remember that VAT is also known as TVA in France.
-"""
-prompt_template = """
-Calculate VAT (aka TVA) for this expense:
-
-@expense
-
-Apply these rates:
-- Restaurant meals: 10% TVA
-- Takeaway/supermarket: 5.5% TVA
-- Alcoholic beverages: 20% TVA
-- Other expenses: Standard rate
-
-Rules:
-1. For amounts < 150€: Ticket de caisse is sufficient if it shows HT, TVA, and TTC
-2. For amounts > 150€: Must have a proper invoice addressed to the company
-3. Multiple TVA rates must be clearly separated on the invoice
-
-Calculate TVA amount and specify rate used.
-"""
