+++ pipelex_libraries/pipelines/examples/extract_generic/__init__.py
@@ -1,0 +1,1 @@
+
