+++ pipelex_libraries/pipelines/examples/extract_generic/extract_generic.toml
@@ -1,0 +1,31 @@
+domain = "extract_generic"
+
+[pipe]
+[pipe.power_extractor]
+PipeSequence = "Update page content with markdown"
+inputs = { page_content = "documents.PageContent" }
+output = "native.Text"
+steps = [
+    { pipe = "extract_page_contents_and_views_from_pdf", result = "page_contents" },  # Located in the base library
+    { pipe = "write_markdown_from_page_content", batch_as = "page_content", result = "markdowns" },
+]
+
+[pipe.write_markdown_from_page_content]
+PipeLLM = "Write markdown from page content"
+inputs = { page_content = "documents.PageContent" }
+output = "native.Text"
+images = ["page_content.page_view"]
+llm = "llm_for_img_to_text"
+system_prompt = """You are a multimodal LLM, expert at converting images into perfect markdown."""
+prompt_template = """
+You are given an image which is a view of a document page.
+You are also given the text extracted from the page by an OCR model.
+Your task is to output the perfect markdown of the page.
+
+Here is the text extracted from the page:
+@page_content.text_and_images.text.text
+
+- Ensure you do not miss any information from the page. The text extracted from the page is not always complete. Your task is to complete the text and add the missing information using the page view.
+- Output only the markdown, nothing else. No need for "```markdown" or "```".
+- In case of diagrams, charts, visualizations, etc. with text inside (the text may not appear in the input text - the OCR fails to extract it), make sure to include the text in the markdown. Feel free to choose the most appropriate markdown element to do so.
+"""
