+++ pipelex_libraries/pipelines/examples/extract_table/table.py
