+++ pipelex_libraries/pipelines/examples/invoice_extractor/__init__.py
@@ -1,0 +1,1 @@
+
