--- pipelex_libraries/pipelines/examples/power_extractor/power_extractor.toml
@@ -1,105 +1,0 @@
-domain = "power_extractor"
-definition = "The domain for power extractor"
-
-[concept.Dpe]
-Concept = "A diagnostic of the energy performance of a building"
-structure = "Dpe"
-
-[concept.ProofOfPurchase]
-Concept = "Elements from a proof of purchase"
-structure = "ProofOfPurchase"
-
-[pipe.write_markdown_from_page_content]
-PipeLLM = "Write markdown from page content"
-inputs = { page_content = "documents.PageContent" }
-output = "TextContent"
-images = ["page_content.page_view"]
-llm = "llm_for_img_to_text"
-system_prompt = """You are a multimodal LLM, expert at converting images into perfect markdown."""
-prompt_template = """
-You are given an image which is a view of a document page.
-You are also given the text extracted from the page by an OCR model.
-Your task is to output the perfect markdown of the page.
-
-Here is the text extracted from the page:
-@page_content.text_and_images.text
-
-- Ensure you do not miss any information from the page. The text extracted from the page is not always complete. Your task is to complete the text and add the missing information using the page view.
-- Output only the markdown, nothing else. No need for "```markdown" or "```".
-- In case of diagrams, charts, visualizations, etc. with text inside (the text may not appear in the input text - the OCR fails to extract it), make sure to include the text in the markdown. Feel free to choose the most appropriate markdown element to do so.
-"""
-
-[pipe.power_extractor]
-PipeSequence = "Update page content with markdown"
-inputs = { page_content = "documents.PageContent" }
-output = "TextContent"
-steps = [
-    { pipe = "extract_page_contents_and_views_from_pdf", result = "page_contents" },
-    { pipe = "write_markdown_from_page_content", batch_as = "page_content", result = "markdowns" },
-]
-
-[pipe.write_markdown_from_page_content_proof_of_purchase]
-PipeLLM = "Write markdown from page content"
-inputs = { page_content = "documents.PageContent" }
-output = "ProofOfPurchase"
-images = ["page_content.page_view"]
-llm = "llm_for_img_to_text"
-structuring_method = "preliminary_text"
-system_prompt = """You are a multimodal LLM, expert at converting images into perfect markdown."""
-prompt_template = """
-You are given an image of a proof of purchase.
-Your role is to convert the image into perfect markdown.
-
-To help you do so, you are given the text extracted from the page by an OCR model.
-@page_content.text_and_images.text
-
-- Ensure you collect every title, number, and currency from the proof of purchase.
-- Pay attention to the text alignment, it might have been misaligned by the OCR.
-- The OCR extraction may be highly incomplete. It is your job to complete the text and add the missing information using the image.
-- Output only the markdown, nothing else. No need for "```markdown" or "```".
-- You can use HTML if it helps you.
-- You can use tables if it is relevant.
-"""
-
-[pipe.power_extractor_proof_of_purchase]
-PipeSequence = "Update page content with markdown"
-inputs = { page_content = "documents.PageContent" }
-output = "TextContent"
-steps = [
-    { pipe = "extract_page_contents_and_views_from_pdf", result = "page_contents" },
-    { pipe = "write_markdown_from_page_content_proof_of_purchase", batch_as = "page_content", result = "proof_of_purchase" },
-]
-
-[pipe.write_markdown_from_page_content_dpe]
-PipeLLM = "Write markdown from page content of a 'Diagnostic de Performance Energetique'"
-inputs = { page_content = "documents.PageContent" }
-output = "Dpe"
-images = ["page_content.page_view"]
-llm = "llm_for_img_to_text"
-structuring_method = "preliminary_text"
-system_prompt = """You are a multimodal LLM, expert at converting images into perfect markdown."""
-prompt_template = """
-You are given an image of a French 'Diagnostic de Performance Energetique'.
-Your role is to convert the image into perfect markdown.
-
-To help you do so, you are given the text extracted from the page by an OCR model.
-@page_content.text_and_images.text
-
-- It is very important that you collect every element, especially if they are related to the energy performance of the building.
-- Pay attention to all the pieces of information that may be included in images, graphs, charts, or tables.
-- We value letters like "A, B, C, D, E, F, G" as they are energy performance classes.
-- Pay attention to the text alignment, it might have been misaligned by the OCR.
-- The OCR extraction may be highly incomplete. It is your job to complete the text and add the missing information using the image.
-- Output only the markdown, nothing else. No need for "```markdown" or "```".
-- You can use HTML if it helps you.
-- You can use tables if it is relevant.
-"""
-
-[pipe.power_extractor_dpe]
-PipeSequence = "Update page content with markdown"
-inputs = { page_content = "documents.PageContent" }
-output = "TextContent"
-steps = [
-    { pipe = "extract_page_contents_and_views_from_pdf", result = "page_contents" },
-    { pipe = "write_markdown_from_page_content_dpe", batch_as = "page_content", result = "dpe" },
-]
