+++ pipelex_libraries/pipelines/examples/write_tweet/tech_tweet.toml
@@ -8,7 +8,6 @@ TweetAnalysis = "Analysis of the tweet's structure and potential improvements"
 WritingStyle = "A style of writing"
 
 [pipe]
-
 [pipe.analyze_tweet]
 PipeLLM = "Analyze the draft tweet and identify areas for improvement"
 inputs = { draft_tweet = "DraftTweet" }
