+++ pipelex_libraries/templates/base_templates.toml
@@ -1,5 +1,3 @@
-
-
 [generic_prompts]
 
 structure_from_preliminary_text_system = "You are a data modeling expert specialized in extracting structure from text."
@@ -13,5 +11,4 @@ Do not create information that is not in the text.
 """
 
 [test_prompts]
-
 jinja2_test_template = "I want a {{ place_holder }} t-shirt."
